# Use rb_str_buf_new() to improve JSON.parse() performance with short string