@Watson1978 2020-11-30
The change with the same effect was made at https://github.com/ruby/ruby/commit/02c32b2e9299726e20bb4499fa7c1f5430ce0b6d

However, this PR is still useful for Ruby 2.7.x or below.