@Watson1978 2021-04-16
Seems this PR isn't necessary any more by https://github.com/flori/json/pull/451/